+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement/Business/Dashboard/Queries/DashboardMappings.cs
@@ -40,7 +40,7 @@ public class GetDocumentsMappings : Profile
         public int Resolve(OutgoingDocument source, GetDocumentResponse destination, int destMember, ResolutionContext context)
         {
             return
-                source.WorkflowHistory
+                (int)source.WorkflowHistory
                     .OrderByDescending(c => c.CreationDate)
                     .FirstDynamic()
                     .Status;
@@ -60,7 +60,7 @@ public class GetDocumentsMappings : Profile
         public int Resolve(IncomingDocument source, GetDocumentResponse destination, int destMember, ResolutionContext context)
         {
             return
-                source.WorkflowHistory
+                (int)source.WorkflowHistory
                     .OrderByDescending(c => c.CreationDate)
                     .FirstDynamic()
                     .Status;
